src/requirements.txt
fastapi
uvicorn
websockets>=10.0
lark-oapi
python-dotenv
openai
requests
redis
pydantic-settings
docker
aiohttp
duckduckgo-search
beautifulsoup4
